Move date rendering to Jinja filter
